Convert AltitudeValidator to single-file app (src/AltitudeBackfill.cs)
Matches the pattern of SampleDataSeeder.cs and VideoSizeBackfill.cs — a
top-level .cs file in src/ with #:project directives, run via
'dotnet run ./src/AltitudeBackfill.cs'. Drops the separate csproj project
and solution entry.
