--- src/Recollections.Tools.AltitudeValidator/Recollections.Tools.AltitudeValidator.csproj
@@ -1,16 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <RootNamespace>Neptuo.Recollections</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Recollections.Data.Ef\Recollections.Data.Ef.csproj" />
-    <ProjectReference Include="..\Recollections.Entries.Azure\Recollections.Entries.Azure.csproj" />
-    <ProjectReference Include="..\Recollections.Entries.SystemIo\Recollections.Entries.SystemIo.csproj" />
-    <ProjectReference Include="..\Recollections.Entries.Data\Recollections.Entries.Data.csproj" />
-    <ProjectReference Include="..\Recollections.Entries\Recollections.Entries.csproj" />
-  </ItemGroup>
-
-</Project>
